% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_multiple_fxout_function.R
\name{read_multiple_fxout}
\alias{read_multiple_fxout}
\title{FoldX Reader Package}
\usage{
read_multiple_fxout(multiple_path, visualize = TRUE)
# read multiple fxout files
}
\arguments{
\item{path}{path to multiple files with .fxout extension.}

\item{visualize}{if TRUE, you will get a visualization of Total energy distribution of all fxout files.}
}
\value{
dataframe of all foldx output exist in that path
}
\description{
The foldX.reader package provides a toolkit for seamlessly importing data from files with a .fxout extension into R, facilitating the creation of database for efficient data management and analysis. for multiple inputs, this package streamlines the process of parsing, organizing, and storing .fxout files into a structured database format and visualize the input. this package has been tested for foldx version 5.
}
\details{
\tabular{lll}{
   \strong{Title} \tab \strong{energy_description} \tab \strong{access} \cr
   \href{https://foldxsuite.crg.eu/energy/TotalEnergy}{Total Energy} \tab This is the predicted overall stability of your protein \tab \code{data.Total_Energy} \cr
   \href{https://foldxsuite.crg.eu/energy/BackboneHbond}{Backbone Hbond} \tab This the contribution of backbone Hbonds \tab \code{data.Backbone_Hbond} \cr
   \href{https://foldxsuite.crg.eu/energy/SidechainHbond}{Sidechain Hbond} \tab This the contribution of sidechain-sidechain and sidechain-backbone Hbonds \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/VanderWaals}{Van der Waals} \tab Contribution of the VanderWaals \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/Electrostatics}{Electrostatics} \tab Electrostatic interactions \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/SolvationPolar}{Solvation Polar} \tab Penalization for burying polar groups \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/SolvationHydrophobic}{Solvation Hydrophobic} \tab Contribution of hydrophobic groups \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/VanderWaalsclashes}{Van der Waals clashes} \tab Energy penalization due to VanderWaals' clashes (interresidue) \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/EntropySideChain}{Entropy Side Chain} \tab Entropy cost of fixing the side chain \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/EntropyMainChain}{Entropy Main Chain} \tab Entropy cost of fixing the main chain \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/SloopEntropy}{Sloop Entropy} \tab ONLY FOR ADVANCED USERS \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/MloopEntropy}{Mloop Entropy} \tab ONLY FOR ADVANCED USERS \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/CisBond}{Cis Bond} \tab Cost of having a cis peptide bond \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/TorsionalClash}{Torsional Clash} \tab VanderWaals' torsional clashes (intraresidue) \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/BackboneClash}{Backbone Clash} \tab Backbone-backbone VanderWaals. These are not considered in the total \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/HelixDipole}{Helix Dipole} \tab Electrostatic contribution of the helix dipole \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/WaterBridge}{Water Bridge} \tab Contribution of water bridges \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/Disulfide}{Disulfide} \tab Contribution of disulfide bonds \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/ElectrostaticKon}{Electrostatic Kon} \tab Electrostatic interaction between molecules in the precomplex \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energyPartialCovalentBonds}{Partial Covalent Bonds} \tab Interactions with bound metals \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/EnergyIonisation}{Energy Ionisation} \tab Contribution of ionisation energy \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/EntropyComplex}{Entropy Complex} \tab Entropy cost of forming a complex \tab \verb{code block} \cr
   \href{https://foldxsuite.crg.eu/energy/ResidueNumber}{Residue Number} \tab Number of residues \tab \verb{code block} \cr
}
}
\references{
FoldX is a molecular modeling software primarily used for protein structure prediction, protein design, and protein engineering studies. It's specifically focused on studying the structure-function relationship of proteins. The software employs empirical force fields and algorithms to predict the stability changes in proteins caused by mutations, analyze protein-protein interactions, and perform protein design experiments. The FoldX Suite builds on the strong fundament of advanced protein design features, already implemented in the successful FoldX3, and exploits the power of fragment libraries, by integrating in silico digested backbone protein fragments of different lengths. Such fragment-based strategy allows for new powerful capabilities: loop reconstruction, implemented in LoopX and peptide docking, implemented in PepX. The Suite also features an improved usability, thanks to a new boost Command Line Interface.

\href{https://foldxsuite.crg.eu/}{FoldX website}
}
\seealso{
\code{\link[=read_fxout]{read_fxout()}}
}
\author{
Kambiz Kalhor
}
